% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bootstrap_exp}
\alias{aus_bootstrap_exp}
\title{Bootstrapping for exponential distribution}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_bootstrap_exp
}
\description{
Bootstrapping for exponential distribution
}
\keyword{datasets}
