% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bootstrap_exp_xmin}
\alias{aus_bootstrap_exp_xmin}
\title{Bootstrapping for exponential distribution with xmin}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_bootstrap_exp_xmin
}
\description{
Bootstrapping for exponential distribution with xmin
}
\keyword{datasets}
