% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bootstrap_pl}
\alias{aus_bootstrap_pl}
\title{Bootstrapping for power law distribution}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_bootstrap_pl
}
\description{
Bootstrapping for power law distribution
}
\keyword{datasets}
