% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_bootstrap_pois}
\alias{aus_bootstrap_pois}
\title{Bootstrapping for poisson distribution}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_bootstrap_pois
}
\description{
Bootstrapping for poisson distribution
}
\keyword{datasets}
