% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_disexp_xmin}
\alias{aus_disexp_xmin}
\title{Exponential distribution with xmin}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_disexp_xmin
}
\description{
Exponential distribution with xmin
}
\keyword{datasets}
