% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_dislnorm}
\alias{aus_dislnorm}
\title{Lognormal distribution}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_dislnorm
}
\description{
Lognormal distribution
}
\keyword{datasets}
