% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_dislnorm_xmin}
\alias{aus_dislnorm_xmin}
\title{Lognormal distribution with xmin}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_dislnorm_xmin
}
\description{
Lognormal distribution with xmin
}
\keyword{datasets}
