% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_displ_xmin}
\alias{aus_displ_xmin}
\title{Power law distribution with xmin}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_displ_xmin
}
\description{
Power law distribution with xmin
}
\keyword{datasets}
