% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_dispois}
\alias{aus_dispois}
\title{Poisson distribution}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_dispois
}
\description{
Poisson distribution
}
\keyword{datasets}
