% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_dispois_xmin}
\alias{aus_dispois_xmin}
\title{Poisson distribution with xmin}
\format{
An object of class \code{list} of length 166.
}
\usage{
aus_dispois_xmin
}
\description{
Poisson distribution with xmin
}
\keyword{datasets}
