% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_frequencies}
\alias{aus_frequencies}
\title{Phoneme frequencies in Australian languages}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4070 rows and 7 columns.
}
\usage{
aus_frequencies
}
\description{
Phoneme frequencies in Australian languages
}
\keyword{datasets}
