% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonfreq_viewer.R
\name{phonfreq_viewer}
\alias{phonfreq_viewer}
\title{Run Phoneme Frequencies Viewer}
\usage{
phonfreq_viewer()
}
\description{
This is a simple wrapper function for opening the phonfreq
viewer app using Shiny. The app produces plots of phoneme frequencies in
Australian languages, plots the fit of different distributions to these
frequencies, and plots the range of uncertainty for various parameter
estimates for each distribution.
}
\details{
A Shiny app for viewing distributions of phoneme frequencies in Australian languages
}
